fetching the basic infos from the server

+++ app/src/main/java/bd/com/ipay/ipayskeleton/Model/MMModule/Resource/District.java
@@ -1,19 +1,19 @@
 package bd.com.ipay.ipayskeleton.Model.MMModule.Resource;
 
 public class District {
-    private long id;
+    private int id;
     private String name;
 
-    public District(long id, String name) {
+    public District(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
